Test save member process

+++ src/main/java/lk/ijse/cmjd108/LibMgmt/controller/MemberController.java
@@ -27,6 +27,9 @@ public class MemberController {
 
     @DeleteMapping
     public ResponseEntity<Void> deleteMember(@RequestParam("memberId") String memberId){
+        if(memberId == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         memberService.deleteMember(memberId);
         return ResponseEntity.noContent().build();
     }
